editing mesh3d shape
git-svn-id: http://b035a0a.online-server.cloud/svn/ITDevelopment@1368 21205d9f-8918-2240-a3d5-c39064a2c852

+++ Include/imt3dgui/CShape3dBase.cpp
@@ -234,7 +234,7 @@ QPoint CShape3dBase::ModelToWindow(const QVector3D& modelCoordinate, const QRect
 }
 
 
-QVector3D CShape3dBase::WindowToModel(const QPoint& windowCoordinate, int z, const QRect& viewPort) const
+QVector3D CShape3dBase::WindowToModel(const QPoint& windowCoordinate, float z, const QRect& viewPort) const
 {
 	if (!m_cameraPtr){
 		return QVector3D();
